// src/model/MenuColour.ts
/**
 * MenuColour is used to define what colour the hamburger menu
 * is based on what page the user's on. If the background colour in
 * the top right part of the screen is light, the menu will be dark.
 * If the background colour is light, the menu will be dark.
 */
export enum MenuColour {
  DARK = '#19062a',
  LIGHT = '#e5e1bd'
}